Move from/to js value to idb implementation itself
