+++ wallet/core/src/storage/transaction/record.rs
@@ -6,7 +6,6 @@ use super::*;
 use crate::imports::*;
 use crate::storage::Binding;
 use crate::tx::PendingTransactionInner;
-use js_sys::{Date, Uint8Array};
 use workflow_core::time::{unixtime_as_millis_u64, unixtime_to_locale_string};
 
 pub use kaspa_consensus_core::tx::TransactionId;
@@ -471,58 +470,6 @@ impl TransactionRecord {
             metadata: None,
             note: None,
         })
-    }
-
-    pub fn to_js_value(&self, secret: Option<&Secret>, encryption_kind: EncryptionKind) -> Result<JsValue, Error> {
-        let id = self.id.to_string();
-        let unixtime_msec = self.unixtime_msec;
-        let mut borsh_data = vec![];
-        <TransactionRecord as BorshSerialize>::serialize(self, &mut borsh_data)?;
-
-        let id_js_value = JsValue::from_str(&id);
-        let timestamp_js_value = match unixtime_msec {
-            Some(unixtime_msec) => {
-                let unixtime_sec = (unixtime_msec / 1000) as u32;
-
-                let date = Date::new_0();
-                date.set_utc_seconds(unixtime_sec);
-                date.into()
-            }
-            None => JsValue::NULL,
-        };
-
-        let encryped_data = if let Some(secret) = secret {
-            Encryptable::from(self.clone()).into_encrypted(secret, encryption_kind)?
-        } else {
-            Encryptable::from(self.clone())
-        };
-        let encryped_data_vec = encryped_data.try_to_vec()?;
-        let borsh_data_uint8_arr = Uint8Array::from(encryped_data_vec.as_slice());
-        let borsh_data_js_value = borsh_data_uint8_arr.into();
-
-        let obj = Object::new();
-        obj.set("id", &id_js_value)?;
-        obj.set("timestamp", &timestamp_js_value)?;
-        obj.set("data", &borsh_data_js_value)?;
-
-        let value = JsValue::from(obj);
-        Ok(value)
-    }
-
-    pub fn from_js_value(js_value: &JsValue, secret: Option<&Secret>) -> Result<Self, Error> {
-        if let Some(object) = Object::try_from(&js_value) {
-            let borsh_data_jsv = object.get_value("data")?;
-            let borsh_data = borsh_data_jsv
-                .try_as_vec_u8()
-                .map_err(|err| Error::Custom(format!("failed to get blob from transaction record object: {:?}", err)))?;
-
-            let encryptable = Encryptable::<TransactionRecord>::try_from_slice(borsh_data.as_slice())?;
-            let transaction_record = encryptable.decrypt(secret)?;
-
-            Ok(transaction_record.0)
-        } else {
-            Err(Error::Custom("supplied argument must be an object".to_string()))
-        }
     }
 }
